Better organizing specs and add capybara
