+++ spec/spec_helper.rb
@@ -1,2 +1,3 @@
 RSpec.configure do |config|
 end
+
