+++ spec/support/controllers/fake_authentication.rb
@@ -1,0 +1,13 @@
+module Controllers
+  module FakeAuthentication
+    def login
+      let!(:login_user) { FactoryGirl.create :user }
+
+      before(:each) do
+        @request.env["devise.mapping"] = Devise.mappings[:user]
+
+        sign_in login_user
+      end
+    end
+  end
+end
