--- spec/support/fake_authentication.rb
@@ -1,11 +1,0 @@
-module FakeAuthentication
-  def login
-    let!(:login_user) { FactoryGirl.create :user }
-
-    before(:each) do
-      @request.env["devise.mapping"] = Devise.mappings[:user]
-
-      sign_in login_user
-    end
-  end
-end
