+++ spec/support/features.rb
@@ -1,0 +1,5 @@
+module Features
+  RSpec.configure do |config|
+    config.include Features::SessionHelpers, type: :feature
+  end
+end
