+++ spec/support/models.rb
@@ -1,0 +1,2 @@
+module Models
+end
